feat: implement profile update functionality
This commit introduces the functionality for users to update their profiles.

The changes include:

- Added a `updateProfile` method in `ProfileController` to handle PUT requests for profiles.
- Implemented `updateProfile` in `ProfileService` to:
    - Verify the profile exists.
    - Verify the handle is not already in use by another profile.
    - Verify the specialities and genres exists.
    - Update the profile in the database.
- Added a `update` method in `ProfileRepository` to update the profile.
- Added `findManyByIds` methods in `GenreRepository` and `SpecialityRepository` to find many genres and specialities by ids.
- Updated `UpdateProfileBodyDTO` to add validation rules and optional fields.

+++ src/modules/profiles/services/profile.service.ts
@@ -15,6 +15,7 @@ import { RequestUserPayloadDTO } from '../dtos/request_user_payload.dto';
 import { AuthService } from '@modules/auth/services/auth.service';
 import { join } from 'path';
 import { unlink } from 'fs/promises';
+import { UpdateProfileBodyDTO } from '../dtos/update_profile_body.dto';
 
 @Injectable()
 export class ProfileService {
@@ -73,22 +74,65 @@ export class ProfileService {
     }
   }
 
-  // async update(profileId: number, profile: UpdateProfileBodyDTO) {
-  //   Logger.log('Updating profile', 'ProfileService');
-  //   const profileToUpdate = await this.profileRepository.findById(profileId);
-  //   if (!profileToUpdate) {
-  //     Logger.error('Profile not found', 'ProfileService');
-  //     throw new AppException({
-  //       message: 'profile not found',
-  //       statusCode: 404,
-  //     });
-  //   }
+  async update(profileId: number, profile: UpdateProfileBodyDTO) {
+    Logger.log('Updating profile', 'ProfileService');
+    const profileToUpdate = await this.profileRepository.findById(profileId);
+    if (!profileToUpdate) {
+      Logger.error('Profile not found', 'ProfileService');
+      throw new AppException({
+        message: 'profile not found',
+        statusCode: 404,
+      });
+    }
+    const handleFinded = await this.profileRepository.findByHandle(
+      profile.handle,
+    );
 
-  //   return await this.profileRepository.update(profileId, {
-  //     ...profile,
-  //     genres: profile.genres,
-  //   });
-  // }
+    if (handleFinded && handleFinded.id !== profileToUpdate.id) {
+      Logger.error('Handle already exists', 'ProfileService');
+      throw new AppException({
+        message: 'Nome de usuário já existe',
+        statusCode: 400,
+      });
+    }
+    if (profile.specialities?.length > 0) {
+      const specialities = await this.specialityRepository.findManyByIds(
+        profile.specialities,
+      );
+      if (specialities.length !== profile.specialities.length) {
+        Logger.error('Specialty not found', 'ProfileService');
+        throw new AppException({
+          message: 'Especialidade não encontrada',
+          statusCode: 404,
+        });
+      }
+      const invalidSpecialities = specialities.filter(
+        (speciality) =>
+          speciality.profile_type_id !== profileToUpdate.profile_type_id,
+      );
+
+      if (invalidSpecialities.length > 0) {
+        Logger.error('Invalid specialities for profile type', 'ProfileService');
+        throw new AppException({
+          message: 'Especialidades inválidas para o tipo de perfil',
+          statusCode: 400,
+        });
+      }
+    }
+
+    if (profile.genres?.length > 0) {
+      const genres = await this.genreRepository.findManyByIds(profile.genres);
+      if (genres.length !== profile.genres.length) {
+        Logger.error('Genre not found', 'ProfileService');
+        throw new AppException({
+          message: 'Gênero não encontrado',
+          statusCode: 404,
+        });
+      }
+    }
+
+    return await this.profileRepository.update(profileId, profile);
+  }
 
   async createOnlyProfile(profile: {
     name: string;
@@ -489,25 +533,14 @@ export class ProfileService {
   }
 
   async addAvatarToProfile(file: { filename: string }, profileId: number) {
-    Logger.log('Adding avatar to profile', 'ProfileService');
-    const profile = await this.profileRepository.findById(profileId);
-    if (!profile) {
-      Logger.error('Profile not found', 'ProfileService');
+    if (!file) {
       throw new AppException({
         error: 'Not found',
-        message: 'profile not found',
+        message: 'file not found',
         statusCode: 404,
       });
     }
-    const avatar = await this.profileRepository.addAvatar(
-      profileId,
-      file.filename,
-    );
-    return avatar;
-  }
-
-  async removeAvatarFromProfile(profileId: number) {
-    Logger.log('Removing avatar from profile', 'ProfileService');
+    Logger.log('Adding avatar to profile', 'ProfileService');
     const profile = await this.profileRepository.findById(profileId);
     if (!profile) {
       Logger.error('Profile not found', 'ProfileService');
@@ -533,6 +566,46 @@ export class ProfileService {
           statusCode: 500,
         });
       }
+    }
+
+    const avatar = await this.profileRepository.addAvatar(
+      profileId,
+      file.filename,
+    );
+    return avatar;
+  }
+
+  async removeAvatarFile(avatar: string) {
+    Logger.log('Removing avatar from profile', 'ProfileService');
+    const filePath = join(process.cwd(), 'uploads', avatar);
+    try {
+      await unlink(filePath);
+    } catch (err) {
+      Logger.warn(
+        `Erro ao remover arquivo de avatar: ${err instanceof Error ? err.message : String(err)}`,
+      );
+      throw new AppException({
+        error: 'Internal Server Error',
+        message: 'Erro ao remover arquivo de avatar',
+        statusCode: 500,
+      });
+    }
+  }
+
+  async removeAvatarFromProfile(profileId: number) {
+    Logger.log('Removing avatar from profile', 'ProfileService');
+    const profile = await this.profileRepository.findById(profileId);
+    if (!profile) {
+      Logger.error('Profile not found', 'ProfileService');
+      throw new AppException({
+        error: 'Not found',
+        message: 'profile not found',
+        statusCode: 404,
+      });
+    }
+
+    if (profile.avatar) {
+      await this.removeAvatarFile(profile.avatar);
     }
 
     await this.profileRepository.removeAvatar(profileId);
@@ -731,6 +804,10 @@ export class ProfileService {
         statusCode: 401,
       });
     }
+    if (profile.avatar) {
+      await this.removeAvatarFile(profile.avatar);
+    }
+
     await this.profileRepository.deleteProfile(profileId);
   }
 }
